requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
billiard==3.3.0.23
black==20.8b1
cachetools==4.2.1
celery==3.1.18
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
Django==3.1.7
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
kombu==3.0.37
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
protobuf==3.15.6
psycopg2==2.7.7
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==2.10.3
regex==2021.3.17
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
